% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crp_related_functions.R
\name{alpha_estimator}
\alias{alpha_estimator}
\title{alpha_estimator}
\usage{
alpha_estimator(crp, alpha_lim, nbre_alpha = 10000)
}
\arguments{
\item{crp}{a numerical vector, the output of \code{\link{rcrp}}.}

\item{alpha_lim}{a two-vector, the range of the potential values for the estimate.}

\item{nbre_alpha}{an interger, the number of the potential values for the estimate.}
}
\value{
A list containing
\describe{
\item{alpha_hat}{the ML estimate,}
\item{elln_alpha}{a useful function to derive the estimate.}
}
}
\description{
Compute the Maximum Likelihood estimate of \code{alpha} from a
\code{crp}.
}
\examples{
crp <- rcrp(5,15)
print(crp)
alpha_ML <- alpha_estimator(crp,c(0.1,10))
alpha_ML$alpha_hat

plot(seq(0.1,10,le=1e4),alpha_ML$elln_alpha,type="l",xlab="",ylab="")
abline(h=length(crp),col="gray")
abline(v=alpha_ML$alpha_hat,col="gray")
}
