% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crp_related_functions.R
\name{augm_rcrp}
\alias{augm_rcrp}
\title{augm_rcrp}
\usage{
augm_rcrp(crp, alpha, N_tilde)
}
\arguments{
\item{crp}{a numerical vector, the output of \code{\link{rcrp}}.}

\item{alpha}{a nonnegative value, the CRP parameter (could be different
from the value used to generate \code{crp}).}

\item{N_tilde}{a nonnegative integer, the number of additional customers.}
}
\value{
A vector of table sizes.
}
\description{
Augment a CRP.
}
\details{
Prefer to use \code{augm_crp_size_cpp} (or \code{augm_crp_assign_cpp}).
}
\examples{
crp <- rcrp(5,15)
print(crp)
print(augm_rcrp(crp,5,0.6))
print(augm_rcrp(crp,5,0.6))
}
