% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Post_Sampler_functions.R
\name{CI_covering}
\alias{CI_covering}
\title{CI_covering}
\usage{
CI_covering(posterior_sample, truth)
}
\arguments{
\item{posterior_sample}{a list, the trace of the sampler algorithm.}

\item{truth}{a list, the true values of the model parameters (see \code{\link{truth1}}).}
}
\value{
A list of boolean indicating if the credible intervals covers the
true values.
}
\description{
Check if the true parameter values belong to the credible intervals.
}
\examples{
res_PT <- PT(param1,data1,hyperparam1)
CI_covering(res_PT$trace,truth1)
}
