% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Post_Sampler_functions.R
\name{compute_CI}
\alias{compute_CI}
\title{compute_CI}
\usage{
compute_CI(posterior_sample, prob = 0.95, chain = 1)
}
\arguments{
\item{posterior_sample}{a list, the trace of the sampler algorithm.}

\item{prob}{a nonnegative value, the coverage probability of the credible interval.}
}
\value{
A list containing the credible intervals.
}
\description{
Compute the credible interval for each model parameters.
}
\examples{
res_PT <- PT(param1,data1,hyperparam1)
compute_CI(res_PT$trace)
}
