% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_for_Sampler.R
\name{compute_starting_point}
\alias{compute_starting_point}
\title{compute_starting_point}
\usage{
compute_starting_point(data, hyperparam)
}
\arguments{
\item{data}{see the example object \code{data1}}

\item{hyperparam}{see the example object \code{hyperparam1}}
}
\value{
a list containing the sampled values: \code{alpha}, \code{c_tilde},
\code{K_tilde}, \code{p} and \code{r}.
}
\description{
Compute a starting point for the posterior sampler.
}
\details{
Prefer to use \code{compute_starting_point_cpp}.
}
\examples{
compute_starting_point(data1,hyperparam1)
}
