% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Post_Sampler_functions.R
\name{compute_summaries}
\alias{compute_summaries}
\title{compute_summaries}
\usage{
compute_summaries(posterior_sample, truth, chain = 1)
}
\arguments{
\item{posterior_sample}{a list, the trace of the sampler algorithm.}

\item{truth}{a list, the true values of the model parameters (see \code{\link{truth1}}).}
}
\value{
a numerical vector.
}
\description{
Compute post sampling summaries to analyze the resulting Markov chain.
}
\details{
The summaries are (for each model parameters): the posterior mean,
the posterior variance, the difference between posterior mean and true value
}
\examples{
res_PT <- PT(param1,data1,hyperparam1)
compute_summaries(res_PT$trace,truth1)
}
