% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data1}
\alias{data1}
\title{a list of objects}
\format{a list of data
\describe{
  \item{c}{the table sizes}
  \item{N}{the number of customers}
  \item{K}{the number of tables}
}}
\usage{
data1
}
\description{
A data object for CRP model
}
\keyword{datasets}
