% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{expand_cube}
\alias{expand_cube}
\title{expand_cube}
\usage{
expand_cube(cube, n)
}
\arguments{
\item{cube}{a cube (3D array).}

\item{n}{a numerical value.}
}
\value{
a cube (3D array).
}
\description{
Internal function to expand cube dimensions when it is necessary.
}
