% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{expand_mat}
\alias{expand_mat}
\title{expand_mat}
\usage{
expand_mat(mat, n)
}
\arguments{
\item{mat}{a matrix.}

\item{n}{a numerical value.}
}
\value{
a matrix.
}
\description{
Internal function to expand matrix dimensions when it is necessary.
}
