% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hyperparam1}
\alias{hyperparam1}
\title{A hyperparam object for the prior specification}
\format{a hyperparam object (list)
\describe{
  \item{lambda1}{shape in the Gamma prior of \code{alpha}}
  \item{lambda2}{rate in the Gamma prior of \code{alpha}}
  \item{eta1}{shape in the Beta prior of \code{p}}
  \item{eta2}{shape in the Beta prior of \code{p}}
  \item{pi}{probability of success in the Negative Binomial distribution of \code{r}}
  \item{kappa}{number of failures in the Negative Binomial distribution of \code{r}}
}}
\usage{
hyperparam1
}
\description{
A hyperparam object for the prior specification
}
\keyword{datasets}
