% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_for_Sampler.R
\name{log_dlkh}
\alias{log_dlkh}
\title{log_dlkh}
\usage{
log_dlkh(theta, data, hyperparam, temperature = 1)
}
\arguments{
\item{theta}{a list of parameter values.}

\item{data}{see the example object \code{data1}}

\item{hyperparam}{see the example object \code{hyperparam1}}

\item{temperature}{a nonnegative value, the temperature parameter required to
compute the targer distribution.}
}
\value{
a numerical value.
}
\description{
Compute the tempered log likelihood.
}
\details{
Prefer to use \code{log_dlkh_cpp}.
}
\examples{
theta <- compute_starting_point(data1,hyperparam1)
log_dlkh(theta,data1,hyperparam1,temperature=1)
log_dlkh(theta,data1,hyperparam1,temperature=0.5)
}
