% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_for_Sampler.R
\name{log_dtarget}
\alias{log_dtarget}
\title{log_dtarget}
\usage{
log_dtarget(theta, data, hyperparam)
}
\arguments{
\item{theta}{a list containing the \code{theta} values.}

\item{data}{see the example object \code{data1}}

\item{hyperparam}{see the example object \code{hyperparam1}}
}
\value{
a data.frame containing the log posterior density, the posterior density,
 the log likelihood, the likelihood, the log prior density and the prior density.
}
\description{
For one \code{theta} value, compute the (log) likelihood
posterior density, prior density and the likelihood, according to the target
 distribution (with \code{temperature = 1}).
}
\details{
Prefer to use \code{log_dtarget_cpp}.
}
\examples{
theta <- compute_starting_point(data1,hyperparam1)
log_dtarget(theta,data1,hyperparam1)
}
