% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_for_Sampler.R
\name{log_dtarget_sample}
\alias{log_dtarget_sample}
\title{log_dtarget_sample}
\usage{
log_dtarget_sample(trace, data, hyperparam)
}
\arguments{
\item{trace}{a list of parameter values.}

\item{data}{see the example object \code{data1}}

\item{hyperparam}{see the example object \code{hyperparam1}}
}
\value{
a data.frame containing the log posterior density, the posterior density,
 the log likelihood, the likelihood, the log prior density and the prior density.
}
\description{
For a sample of \code{theta}, compute the (log) likelihood
posterior density, prior density and the likelihood, according to the target
 distribution (with \code{temperature = 1}).
}
\details{
Prefer to use \code{log_dtarget_sample_cpp}.
}
\examples{
res_MH <- MH(param1,data1,hyperparam1,verbose=F)
log_dtarget_sample(res_MH,data1,hyperparam1)
}
