% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sampler.R
\name{MH}
\alias{MH}
\title{Metropolis-Hastings}
\usage{
MH(param, data, hyperparam, verbose = TRUE)
}
\arguments{
\item{param}{a list containing options for the sampler algorithm:
\describe{
  \item{n_iter}{the size of the ouput posterior sample}
  \item{temperature}{(optional) the temperatures used to define the chain invariant distributions}
  \item{lambda}{(optional) related to the random number of swap moves per iteration}
  \item{thin}{(optional) keep only one iteration every \code{thin}-iterations.}
  \item{burnin}{(optional) an integer, drop the first iterations.}
  \item{target}{(optional) a string, "posterior" (default) or "prior".}
}}

\item{data}{see the example object \code{data1}}

\item{hyperparam}{see the example object \code{hyperparam1}}

\item{verbose}{Write stuff.}
}
\value{
a list containing the \code{theta} values and other useful quantities.
}
\description{
A Metropolis-Hastings algorithm to sample from the prior or the
posterior.
}
\details{
Should only be used for obtaining a prior sample: \code{param$target = "prior"}

Prefer to use \code{MH_cpp}.
}
\examples{
param1$target <- "prior"
res_MH <- MH(param1,data1,hyperparam1)
}
