% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{param1}
\alias{param1}
\title{A param object for the posterior sampler}
\format{a parma object (list)
\describe{
  \item{n_iter}{the size of the ouput posterior sample}
  \item{temperature}{the temperatures used to define the chain invariant distributions}
  \item{lambda}{related to the random number of swap moves per iteration}
  \item{thin}{keep only one iteration every \code{thin}-iterations.}
}}
\usage{
param1
}
\description{
A param object for the posterior sampler
}
\keyword{datasets}
