% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloCRP.R
\docType{package}
\name{phyloCRP}
\alias{phyloCRP}
\alias{phyloCRP-package}
\title{phyloCRP: Modelling phylogenetic subtrees with a Chinese Restaurant Process.}
\description{
This is the implementation of a method to fit a phylogenetic subtrees model.
The main idea is set up a statistical model that describes a collection of
phylogenetic subtrees in order to take into account 1) the uncertainty of the
input data and 2) the data incompleteness assumption.
}
