% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sampler.R
\name{PT}
\alias{PT}
\title{Parallele Tempering}
\usage{
PT(param, data, hyperparam, verbose = T)
}
\arguments{
\item{param}{a list containing options for the sampler algorithm:
\describe{
  \item{n_iter}{the size of the ouput posterior sample}
  \item{temperatures}{(optional) the temperatures used to define the chain invariant distributions}
  \item{lambda}{(optional) related to the random number of swap moves per iteration}
  \item{thin}{(optional) keep only one iteration every \code{thin}-iterations.}
  \item{burnin}{(optional) an integer, drop the first iterations.}
  \item{target}{(optional) a string, "posterior" (default) or "prior".}
}}

\item{data}{see the example object \code{data1}}

\item{hyperparam}{see the example object \code{hyperparam1}}

\item{verbose}{Write stuff.}
}
\value{
a list containing the computational time and the list of the \code{MH}
output for each parallel chains (see \code{\link{MH}}).
}
\description{
A Parallel Tempering algorithm to sample from the target distribution.
}
\details{
Prefer to use \code{PT_cpp}.
}
\examples{
res_PT <- PT(param1,data1,hyperparam1)
}
