% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sampler.R
\name{PT_bs}
\alias{PT_bs}
\title{Parallele Tempering (with bootstrap)}
\usage{
PT_bs(param, datas, hyperparam, verbose = TRUE)
}
\arguments{
\item{param}{a list containing options for the sampler algorithm:
\describe{
  \item{n_iter}{the size of the ouput posterior sample}
  \item{temperatures}{(optional) the temperatures used to define the chain invariant distributions}
  \item{lambda}{(optional) related to the random number of swap moves per iteration}
  \item{thin}{(optional) keep only one iteration every \code{thin}-iterations.}
  \item{burnin}{(optional) an integer, drop the first iterations.}
  \item{target}{(optional) a string, "posterior" (default) or "prior".}
}}

\item{datas}{a list of data, each item should be as the example object \code{data1}}

\item{hyperparam}{see the example object \code{hyperparam1}}

\item{verbose}{Write stuff.}
}
\value{
a list  containing the result (for each
 bootstrap chain) of the \code{MH} output for each parallel chains
 (see \code{\link{MH}}).
}
\description{
A Parallel Tempering algorithm to sample from the target distribution,
with bootstrap input data.
}
\details{
Prefer to use \code{PT_bs_cpp}.
}
\examples{
datas <- list(data1,data1,data1)
res_PT <- PT_bs(param1,datas,hyperparam1)
}
