% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{reduce_cube}
\alias{reduce_cube}
\title{reduce_cube}
\usage{
reduce_cube(cube)
}
\arguments{
\item{cube}{a cube (3D array).}
}
\value{
a cube (3D array).
}
\description{
Internal function to reduce cube dimensions when it is necessary.
}
