% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{reduce_mat}
\alias{reduce_mat}
\title{reduce_mat}
\usage{
reduce_mat(mat)
}
\arguments{
\item{mat}{a matrix.}
}
\value{
a matrix.
}
\description{
Internal function to reduce matrix dimensions when it is necessary.
}
