% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_for_Sampler.R
\name{rprior}
\alias{rprior}
\title{rprior}
\usage{
rprior(data, hyperparam)
}
\arguments{
\item{data}{see the example object \code{data1}}

\item{hyperparam}{see the example object \code{hyperparam1}}
}
\value{
a list containing the sampled values: \code{alpha}, \code{c_tilde},
\code{K_tilde}, \code{p} and \code{r}.
}
\description{
Sample from the hierarchical prior distribution#'
}
\details{
Prefer to use \code{rprior_cpp}.
}
\examples{
rprior(data1,hyperparam1)
}
