% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crp_related_functions.R
\name{sub_rcrp}
\alias{sub_rcrp}
\title{sub_rcrp}
\usage{
sub_rcrp(crp, p)
}
\arguments{
\item{crp}{a numerical vector, the output of \code{\link{rcrp}}.}

\item{p}{a numerical value belonging to (0,1), the sampling proportion.}
}
\value{
A vector of table sizes.
}
\description{
Subsample from a CRP with a Binomial model.
}
\details{
Prefer to use \code{subcrp_size_cpp}.
}
\examples{
crp <- rcrp(5,15)
print(crp)
print(sub_rcrp(crp,0.6))
print(sub_rcrp(crp,0.6))
}
