% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{truth1}
\alias{truth1}
\title{A list of the true parameter values}
\format{a list of the true parameter values
\describe{
  \item{alpha}{the value used to generate the true complete table sizes}
  \item{p}{the true sampling proportion}
  \item{K_tilde}{the true number of tables}
  \item{c_tilde}{the true complete table sizes}
  \item{r}{the true number of missing customers}
}}
\usage{
truth1
}
\description{
A list of the true parameter values
}
\keyword{datasets}
